# perf: Improve performance by caching verifier