<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Description

Closes: #358 

This PR fixes an issue causing `fast-jwt` verifier to be recreated on every request, which means the `fast-jwt` cache was not being used. By caching the `fast-jwt` verifier based on the options passed in, an existing verifier instance can be reused to verify incoming request JWTs instead of having to recreate a new verifier, and then the `fast-jwt` cache can be utilized (if enabled).

#### Benchmarks

Tested using a simple fastify server with autocannon used to create load

No auth: 77,978 req/s

Changes (no cache): 53,008 req/s
Changes (cache): 60,880  req/s

Main (no cache): 46,302 req/s
Main (cache): 34,669.1 req/s

![image](https://github.com/user-attachments/assets/c5a9ec5a-9234-462f-b89f-fc53ab082fb2)

> Very odd behaviour on current main where enabling cache has a performance impact, I believe this is caused by fast-jwt having to recreate cache on every request.

Config used:

```ts
fastify.register(require('@fastify/jwt'), {
  secret: 'supersecret',
  decode: { complete: true },
  verify: {
     cache: true, // disabled for "no cache" tests
  },
})
```


#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)