@SeanReece 2025-01-20
@kibertoad Thanks for reaching out - work has just been busy lately. I've been spending a bit more time getting familiar with the `fastify-jwt` source and I think I have a better solution instead of caching verifier instances. Hoping to have something today.